피보호자 설정 컴포넌트 분리
