+++ frontend/src/components/WardSettings/WardSettingsEditForm.js
@@ -1,7 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import api from '../../services/Api';
+import WardCheck from './WardCheck';
 
-const WardSettingsEditForm = () => {
+const WardSettingsEditForm = ({ isEdit, setIsEdit }) => {
     const [nameCheck, setNameCheck] = useState('');
     const [ageCheck, setAgeCheck] = useState('');
     const [wardinfo, setWardInfo] = useState({
@@ -13,8 +14,6 @@ const WardSettingsEditForm = () => {
         remark: '',
     });
 
-    const [isEdit, setIsEdit] = useState(false)
-    const [btn, setBtn] = useState('수정')
     const [photoUpdated, setPhotoUpdated] = useState(false); //사용자의 사진 업데이트 여부
     const [previewUrl, setPreviewUrl] = useState(''); // 미리보기 URL 상태
 
@@ -88,53 +87,6 @@ const WardSettingsEditForm = () => {
         }
     };
 
-    const wardSetting = () => {
-        setIsEdit(!isEdit)
-        setBtn(isEdit ? '수정' : '확인')
-        if (btn === '확인') {
-            infoCheck()
-        }
-    };
-
-    const newAge = parseInt(wardinfo.age);
-    const validAge = newAge >= 1 && newAge <= 100;
-    const validInfo = wardinfo.name && wardinfo.age;
-
-    useEffect(() => {
-        if (wardinfo.name || wardinfo.age) {
-            if (wardinfo.name === '' || wardinfo.age === '') {
-                (wardinfo.name === '') ? setNameCheck('필수항목입니다.') : setAgeCheck('필수항목입니다.')
-            }
-            if (wardinfo.name) {
-                const validNumbers = /\d/.test(wardinfo.name);
-                if (validNumbers) {
-                    setNameCheck('이름은 영어 또는 한글로 입력해 주세요.')
-                }
-                else {
-                   setNameCheck('') 
-                }
-            }
-            if (wardinfo.age) {
-                setAgeCheck(validAge ? '' : '1~100 사이의 정수만 입력해 주세요.')
-            }
-        }
-        else {
-            setNameCheck('필수항목입니다.')
-            setAgeCheck('필수항목입니다.')
-        }
-    }, [wardinfo]);
-
-    const infoCheck = () => {
-        if (validAge && validInfo && nameCheck === '' && ageCheck === '') {
-            saveWardSetting(wardinfo.photo, wardinfo.photo_url, wardinfo.name, wardinfo.age, wardinfo.remark)
-        } 
-        else {
-            alert("입력되지 않은 값이 있습니다.")
-            setIsEdit(true)
-            setBtn('확인')
-        }
-    };
-
     return (
         <div className='flex-col text-left justify-start pl-24 mt-16 mb-6'> {isEdit ? (
             <>
@@ -170,7 +122,10 @@ const WardSettingsEditForm = () => {
                 })}
                 </div>
                 </>}
-            <button className='input-box2 absolute left-[650px] top-[635px] p-2 w-40 text-3xl' onClick={wardSetting}>{btn}</button>
+            <WardCheck wardinfo={wardinfo} nameCheck={nameCheck} setNameCheck={setNameCheck} 
+            ageCheck={ageCheck} setAgeCheck={setAgeCheck} saveWardSetting={saveWardSetting}
+            isEdit={isEdit} setIsEdit={setIsEdit}
+            />
         </div>
     );
 };
